add stack a

+++ libft/libft.h
@@ -40,7 +40,7 @@ char	*ft_itoa(int n);
 char	*ft_strjoin(char const *s1, char const *s2);
 char	*ft_strtrim(char const *s1, char const *set);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
-void	ft_putnbr_fd(int n, int fd);
+void	ft_putnbr_fd(long int n, int fd);
 char	**ft_split(char const *s, char c);
 size_t	ft_count_words(const char *s, char c);
 char	*ft_mystrncpy(char *dst, const char *src, size_t len);
